extract util funs to file
